Move player & unit to babylon dir
